requirements.txt
aiofiles==25.1.0
aiosqlite==0.22.1
asyncua==1.1.8
black==26.1.0
c104==2.2.1
cffi==2.0.0
click==8.3.1
coverage==7.13.1
cryptography==46.0.3
dnp3py==0.1.0
iniconfig==2.3.0
isort==7.0.0
librt==0.7.8
mypy==1.19.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pluggy==1.6.0
pycomm3==1.2.16
pycparser==2.23
Pygments==2.19.2
pymodbus==3.11.4
pyOpenSSL==25.3.0
pyserial==3.5
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-ordering==0.6
python-dateutil==2.9.0.post0
python-snap7==2.0.2
pytokens==0.4.0
pytz==2025.2
PyYAML==6.0.3
ruff==0.14.13
six==1.17.0
sortedcontainers==2.4.0
typing_extensions==4.15.0
